Move file to where other branch moved it, to make merge easier to reason about
